% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptoticComplexityClass.R
\name{asymptoticComplexityClass}
\alias{asymptoticComplexityClass}
\title{Asymptotic Complexity Classification function}
\usage{
asymptoticComplexityClass(df, output.size, data.size)
}
\arguments{
\item{df}{A data frame composing for two columns at the least, where one should be the contain the output-parameter sizes and one should contain the data sizes.}

\item{output.size}{A string specifying the column name in the passed data frame to be used as the output size.}

\item{data.size}{A string specifying the column name in the passed data frame to be used as the data size.}
}
\value{
A string specifying the resultant complexity class. (Eg: 'Linear', 'Log-linear', 'Quadratic')
}
\description{
Function to classify the complexity trend between two selected parameters from the data frame provided as input here
}
