% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_complexity_class.R
\name{expect_complexity_class}
\alias{expect_complexity_class}
\title{Expected Complexity Class test function}
\usage{
expect_complexity_class(object, complexity.class)
}
\arguments{
\item{object}{The argument which returns a string to be tested against the expected string}

\item{complexity.class}{A string denoting the expected complexity class}
}
\value{
null for expected complexity, else throws error.
}
\description{
Function to test if object returning/printing a string matches with a specified string
}
