% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_linear_time.R
\name{expect_linear_time}
\alias{expect_linear_time}
\title{Linear Time Complexity Test function}
\usage{
expect_linear_time(f = asymptoticTimings, ...)
}
\arguments{
\item{f}{A function which returns a data frame composed of timings and corresponding data sizes, ideally from asymptoticTimings. (can use other functions)}

\item{...}{Parameters for passed function 'f'.}
}
\value{
null for expected complexity (linear), else throws error.
}
\description{
Function to test if input algorithm has linear time complexity
}
