% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglinearvsquadraticClassifier.R
\name{loglinearvsquadraticClassifier}
\alias{loglinearvsquadraticClassifier}
\title{Log-linear vs Quadratic time complexity classifier}
\usage{
loglinearvsquadraticClassifier(f1, f2)
}
\arguments{
\item{f1}{Function based on an algorithm which is known or expected to run at either log-linear or quadratic time complexity, different from that of f2.}

\item{f2}{Function based on an algorithm which is known or expected to run at either log-linear or quadratic time complexity, different from that of f1.}
}
\value{
This function does not return any value but a statement indicating which among the two is log-linear and which is quadratic in nature.
}
\description{
Preliminary function to distinguish between log-linear vs quadratic time complexity
of two algorithms/functions with parameters :            count.vec,     <- integer vector of count data.
                              weight.vec=rep(1, length(count.vec)),     <- numeric vector (same length as count.vec) of positive weights.
                                                  and max.segments.     <- integer of length 1: maximum number of segments (must be >= 2, default value: 3).
}
