% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMemoryUsage.R
\name{plotMemoryUsage}
\alias{plotMemoryUsage}
\title{Asymptotic Memory Usage Plot function}
\usage{
plotMemoryUsage(data.df)
}
\arguments{
\item{data.df}{A data frame composed of columns 'Memory Usage' and 'Data sizes', which can be obtained by asymptoticMemoryUsage()}
}
\value{
A ggplot object.
}
\description{
Function to plot timings vs data sizes from the data frame returned by asymptoticMemoryUsage()
}
