% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptoticComplexityClassifier.R
\name{asymptoticComplexityClassifier}
\alias{asymptoticComplexityClassifier}
\title{Asymptotic Complexity Classifier function}
\usage{
asymptoticComplexityClassifier(df)
}
\arguments{
\item{df}{A data frame returned by asymptoticComplexityClass.}
}
\value{
A string specifying the resultant complexity class. (Eg: 'Linear', 'Log-linear', 'Quadratic')
}
\description{
Function to classify the complexity trend between the two parameters as provided by asymptoticComplexityClass
}
\details{
For more information regarding its implementation or functionality/usage, please check https://anirban166.github.io//Generalized-complexity/
}
\examples{
# Quantifying timings for the substring function against a set of increasing input data sizes:
input.sizes = 10^seq(1, 4, by = 0.5)
timings.df <- asymptoticTimings(substring(paste(rep("A", N), collapse = ""), 1:N, 1:N), input.sizes)
# Classifying the complexity trend between the data contained in the columns
# 'Timings' and 'Data sizes' from the data frame obtained above:
df <- data.frame('output' = timings.df$Timings, 'size' = timings.df$`Data sizes`)
asymptoticComplexityClassifier(df)
# Expected complexity class: linear
}
