% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptoticComplexityClassifier.R
\name{asymptoticComplexityClassifier}
\alias{asymptoticComplexityClassifier}
\title{Asymptotic Complexity Classifier function}
\usage{
asymptoticComplexityClassifier(df)
}
\arguments{
\item{df}{A data frame returned by asymptoticComplexityClass.}
}
\value{
A string specifying the resultant complexity class. (Eg: 'Linear', 'Log-linear', 'Quadratic')
}
\description{
Function to classify the complexity trend between the two parameters as provided by asymptoticComplexityClass
}
\details{
For more information regarding its implementation or functionality/usage, please check https://anirban166.github.io//Generalized-complexity/
}
