% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptoticTimeComplexityClass.R
\name{asymptoticTimeComplexityClass}
\alias{asymptoticTimeComplexityClass}
\title{Asymptotic Time Complexity Classifying function}
\usage{
asymptoticTimeComplexityClass(model.df)
}
\arguments{
\item{model.df}{A data frame composed of columns 'Timings' and 'Data sizes', which can be obtained by asymptoticTimings()}
}
\value{
A string specifying the resultant complexity class. (Eg: 'Linear', 'Log-linear', 'Quadratic')
}
\description{
Function to classify the time complexity of the expression passed onto asymptoticTimings (which outputs a data frame to be used here)
}
