% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_linear_time.R
\name{expect_linear_time}
\alias{expect_linear_time}
\title{Linear Time Complexity Test function}
\usage{
expect_linear_time(...)
}
\arguments{
\item{...}{Parameters for the expression to be tested, which is a function of data.sizes, as in asymptoticTimings}
}
\value{
null for expected complexity (linear), else throws error.
}
\description{
Function to test if input algorithm has linear time complexity
}
\details{
For more information regarding its implementation or functionality/usage, please check https://anirban166.github.io//Testing-functions/
}
\examples{
\donttest{
# Avoiding for CRAN since computation time can exceeds 5 seconds in some runs:
# Testing a simple expression:
input.sizes <- 10^seq(1, 4, by = 0.5)
expect_linear_time(N, input.sizes)
# The above code will throw an error if the function does not follow a linear trend.
}
}
