% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_linear_time.R
\name{expect_linear_time}
\alias{expect_linear_time}
\title{Linear Time Complexity Test function}
\usage{
expect_linear_time(...)
}
\arguments{
\item{...}{Parameters for the expression to be tested, which is a function of data.sizes, as in asymptoticTimings}
}
\value{
null for expected complexity (linear), else throws error.
}
\description{
Function to test if input algorithm has linear time complexity
}
