% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_loglinear_time.R,
%   R/expect_quadratic_time.R
\name{expect_loglinear_time}
\alias{expect_loglinear_time}
\title{Log-Linear Time Complexity Test function}
\usage{
expect_loglinear_time(f = asymptoticTimings, ...)

expect_loglinear_time(f = asymptoticTimings, ...)
}
\arguments{
\item{f}{A function which returns a data frame composed of timings and corresponding data sizes, ideally from asymptoticTimings. (can use other functions)}

\item{...}{Parameters for passed function 'f'.}
}
\value{
null, throws error if complexity is not log-linear. (i.e. O(NlogN) for size N, in Big-O notation)

null for expected complexity (quadratic), else throws error.
}
\description{
Function to test if input algorithm has log-linear time complexity

Function to test if input algorithm has quadratic time complexity
}
