% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_quadratic_time.R
\name{expect_quadratic_time}
\alias{expect_quadratic_time}
\title{Quadratic Time Complexity Test function}
\usage{
expect_quadratic_time(...)
}
\arguments{
\item{...}{Parameters for the expression to be tested, which is a function of data.sizes, as in asymptoticTimings}
}
\value{
null for expected complexity (quadratic), else throws error.
}
\description{
Function to test if input algorithm has quadratic time complexity
}
\details{
For more information regarding its implementation or functionality/usage, please check https://anirban166.github.io//Testing-functions/
}
