% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_quadratic_time.R
\name{expect_quadratic_time}
\alias{expect_quadratic_time}
\title{Quadratic Time Complexity Test function}
\usage{
expect_quadratic_time(f = asymptoticTimings, ...)
}
\arguments{
\item{f}{A function which returns a data frame composed of timings and corresponding data sizes, ideally from asymptoticTimings. (can use other functions)}

\item{...}{Parameters for passed function 'f'.}
}
\value{
null for expected complexity (quadratic), else throws error.
}
\description{
Function to test if input algorithm has quadratic time complexity
}
