% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimings.R
\name{plotTimings}
\alias{plotTimings}
\title{Asymptotic Timings Plot function}
\usage{
plotTimings(data.df)
}
\arguments{
\item{data.df}{A data frame composed of columns 'Timings' and 'Data sizes', which can be obtained by asymptoticTimings()}
}
\value{
A ggplot object.
}
\description{
Function to plot timings vs data sizes from the data frame returned by asymptoticTimings
}
