% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimings.R
\name{plotTimings}
\alias{plotTimings}
\title{Asymptotic Timings Plot function}
\usage{
plotTimings(
  data.df,
  titles = list("", ""),
  labels = list("Data size", "Runtime (in nanoseconds)"),
  point.alpha = 1,
  line.alpha = 1,
  point.color = "black",
  line.color = "black",
  point.size = 1.3,
  line.size = 0.7
)
}
\arguments{
\item{data.df}{A data frame composed of columns 'Timings' and 'Data sizes', which can be obtained by asymptoticTimings()}

\item{titles}{A list of two elements consisting of strings for the plot title and subtitle.
# Optional, with default values set to empty strings. (no titles/subtitles)}

\item{labels}{A list of two elements containing strings for x and y labels respectively.
# Optional, with default values set to appropriate labels.}

\item{point.alpha}{A numeric value denoting transparency level (in the range 0 to 1) for point geometry.
# Optional, with the default value set to 1. (no transparentness)}

\item{line.alpha}{A numeric value denoting transparency level (in the range 0 to 1) for line geometry.
# Optional, with the default value set to 1. (no transparentness)}

\item{point.color}{A string specifying a known color or a representation in hexcode for point geometry.
# Optional, with the default color set as black. (Hex equivalent: #000000)}

\item{line.color}{A string specifying a known color or a representation in hexcode for line geometry.
# Optional, with the default color set as black. (Hex equivalent: #000000)}

\item{point.size}{A numeric value denoting the size of point geometry.
# Optional, with the default value set to (1.3).}

\item{line.size}{A numeric value denoting the size of line geometry.
# Optional, with the default value set to (0.7).}
}
\value{
A ggplot object.
}
\description{
Function to plot timings vs data sizes from the data frame returned by asymptoticTimings
}
\details{
For more information regarding its implementation or functionality/usage, please check https://anirban166.github.io//Timings-plotting-function/
}
